Se agregaron los 9 endpoints restantes del apartado de administrador, se tuvieron en cuenta el manejo de errores con el handler, utilizando las excepcion ya existentes, se sacaron los @Service de las interfaces, se comento todo lo innecesario en los distintos archivos modificados y se dejo algunos comentarios para futuras revisiones. En unos de los endpoints se reutilizo un dto para no crear otro con los mismos campos, esto esta a revision tambien.

+++ Poo2025/src/main/java/ar/edu/unnoba/poo2025/torneos/service/TorneoServiceImp.java
@@ -1,11 +1,14 @@
 package ar.edu.unnoba.poo2025.torneos.service;
 
+import java.math.BigDecimal;
+import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import ar.edu.unnoba.poo2025.torneos.dto.ActualizarTorneoDTO;
 import ar.edu.unnoba.poo2025.torneos.dto.TorneoDetalleDTO;
 import ar.edu.unnoba.poo2025.torneos.exception.DuplicateResourceException;
 import ar.edu.unnoba.poo2025.torneos.exception.ResourceNotFoundException;
@@ -24,7 +27,7 @@ public class TorneoServiceImp implements TorneoService {
     public List<TorneoModel> obtenerTorneos() {
         return torneoRepository.findAll();
     }
-
+/* 
     @Override
     public List<TorneoModel> findByPublishedTrue() {
         java.sql.Date hoy = new java.sql.Date(System.currentTimeMillis());
@@ -37,7 +40,29 @@ public class TorneoServiceImp implements TorneoService {
             }
         }
         return aux;
+    }*/
+    @Override
+    public List<TorneoModel> findByPublishedTrue() {
+        LocalDate hoy = LocalDate.now();
+        List<TorneoModel> aux = new ArrayList<>();
+
+        for (TorneoModel torneo : obtenerTorneos()) {
+
+            LocalDate ini = torneo.getFechaIni();
+            LocalDate fin = torneo.getFechaFin();
+
+            boolean estaPublicado = Boolean.TRUE.equals(torneo.getPublicado());
+            boolean iniEsHoyOPasado =ini.isBefore(hoy) || ini.isEqual(hoy);  
+            boolean finEsNuloOAnteriorAHoy =fin == null || fin.isBefore(hoy);        
+
+            if (estaPublicado && iniEsHoyOPasado && finEsNuloOAnteriorAHoy) {
+                aux.add(torneo);
+            }
+        }
+
+        return aux;
     }
+
 
     @Override
     public TorneoModel obtenerPorId(Long id) {
@@ -47,41 +72,45 @@ public class TorneoServiceImp implements TorneoService {
 
     @Override
     public void eliminar(Long id) {
+        if(torneoRepository.findById(id).equals(id)){
+            throw new ResourceNotFoundException("No se encontro el torneo con el ID: "+id);
+        }
         torneoRepository.deleteById(id);
     }
-
+ 
+ 
+    //TP6 
     @Override
-    public void actualizar(Long id, TorneoModel nuevosDatos) {
+    public void actualizar(Long id, ActualizarTorneoDTO nuevosDatos) {
         TorneoModel existente = obtenerPorId(id);
+                
         existente.setNombre(nuevosDatos.getNombre());
         existente.setDescripcion(nuevosDatos.getDescripcion());
         existente.setFechaIni(nuevosDatos.getFechaIni());
         existente.setFechaFin(nuevosDatos.getFechaFin());
-        existente.setPublicado(nuevosDatos.getPublicado());
         torneoRepository.save(existente);
-    }
-
-    //TP6 
+         
+    } 
     @Override
     public List<TorneoModel> getTorneosOrdenadosDesc() {
         List<TorneoModel> aux = new ArrayList<>(obtenerTorneos());
         aux.sort((t1, t2) -> t2.getFechaIni().compareTo(t1.getFechaIni()));
         return aux;
     }
-
+    //revisar el problema del bigdecimal 
     @Override
     public TorneoDetalleDTO obtenerTorneoDetalleDTO(Long torneoId) {
         TorneoModel torneo = torneoRepository.findById(torneoId)
                 .orElseThrow(() -> new ResourceNotFoundException("No se encontró un torneo con el ID: " + torneoId));
-
+ 
         int totalInscripciones = 0;
-        double montoTotal = 0;
+        BigDecimal montoTotal = BigDecimal.ZERO;
         if (torneo.getCompetencias() != null) {
             for (CompetenciaModel c : torneo.getCompetencias()) {
                 if (c.getInscripciones() != null) {
                     totalInscripciones += c.getInscripciones().size();
                     for (InscripcionModel i : c.getInscripciones()) {
-                        montoTotal += i.getPrecioPagado();
+                        montoTotal = montoTotal.add(i.getPrecioPagado());
                     }
                 }
             }
@@ -104,4 +133,11 @@ public class TorneoServiceImp implements TorneoService {
         }
         torneoRepository.save(torneo);
     }
+    @Override
+    public void cambiarEstadoAPublicado(Long idTorneo){
+        TorneoModel torneo = obtenerPorId(idTorneo);
+        torneo.setPublicado(true);
+        torneoRepository.save(torneo);
+    }  
+
 }
